organize pkg `client`
