+++ pkg/client/cmd/app.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 
 	"github.com/fatih/color"
-	"github.com/luizalabs/teresa-api/cmd/client/connection"
 	"github.com/luizalabs/teresa-api/pkg/client"
+	"github.com/luizalabs/teresa-api/pkg/client/connection"
 	appb "github.com/luizalabs/teresa-api/pkg/protobuf/app"
 	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
