+++ pkg/client/cmd/completion.go
