+++ pkg/client/cmd/delete.go
